refactor(team): move team module to nanobot/team, filelock to utils, fix circular imports
