+++ nanobot/team/__init__.py
@@ -1,0 +1,6 @@
+"""Nano Team Mode — LLM-planned multi-worker async collaboration."""
+
+from nanobot.team.manager import TeamManager
+from nanobot.team.state import TeamRuntime, TeamState
+
+__all__ = ["TeamManager", "TeamRuntime", "TeamState"]
