+++ nanobot/utils/__init__.py
@@ -1,5 +1,6 @@
 """Utility functions for nanobot."""
 
+from nanobot.utils.filelock import filelock
 from nanobot.utils.helpers import ensure_dir
 
-__all__ = ["ensure_dir"]
+__all__ = ["ensure_dir", "filelock"]
